refactor(embeddings): extract embeddings-chunks module
